pyproject.toml
[project]
name = "final-year-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.135.1",
    "google-genai>=1.66.0",
    "pillow>=12.1.1",
    "python-multipart>=0.0.22",
    "uvicorn>=0.41.0",
]
